// src/generated/models/userCreate.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * iPhande API
 * Township enterprise ERP, hyper-local marketplace, and merchant credibility infrastructure.
 * OpenAPI spec version: 0.1.0
 */
import type { UserCreateBusinessName } from './userCreateBusinessName';
import type { UserCreateBusinessType } from './userCreateBusinessType';
import type { UserCreateCity } from './userCreateCity';
import type { UserCreateProvince } from './userCreateProvince';

export interface UserCreate {
  email: string;
  password: string;
  role?: string;
  business_name?: UserCreateBusinessName;
  business_type?: UserCreateBusinessType;
  city?: UserCreateCity;
  province?: UserCreateProvince;
}